overhaul migration system
(This commit also removes per-store item codecs. I decided that they should be gone, and figured there's no point in re-implementing them with this overhaul when I'd then immediately remove them. I suppose perhaps I could have done a 'remove item codecs' commit before this one, but I think I made the decision part of the way through this commit. So be it.)

This was tough to design. See, the migration system has several requirements which unfortunately are in friction with each other.

Essentially, the working design has 3 major parts in each migration (version upgrade):
  1. Database alterations
  2. Async work
  3. Type registering

(1.) is the actual changes to the DB--adding and removing stores and indexes
(2.) is arbitrary user-given async work, such as reformatting existing rows to match the alterations
(3.) is registering custom Storable or Indexable types, or re-registering types to match the alterations

Now, this would mostly all be fine. We could just let people do what they want within a versionchange transaction and be done with it.

However, this comes with an issue. The issue is that, at runtime, each index needs to know what trait it's tracking. This information is only given when the index is initially defined, which is part of the datbase alterations in migrations. It would have been nice to be able to serialize and store this information between sessions, but that's not possible since traits can be functions. Thus this information needs to be recalculated every runtime.

Previously, this was not a big problem, since migrations were split into 3 stages: (1) custom async 'before alterations' work; (2) alterations; (3) custom async 'after alterations' work. Every runtime would use the alterations information to recalculate the index trait getters without actually running the alterations or doing the async work.

This approach was somewhat inflexible to begin, since it requires all migrations to be [async work -> alterations -> async work]. But I figured that that's a reasonable requirement. However, with the addition of type registering, this would have become more difficult. Would it be [type registering -> async work -> alterations -> async work -> type registering]? Or would type registering sandwich alterations on the inside? It's unclear. Or perhaps type registering could just be a part of async work? But IIRC doing this would have produced a rather tricky API or something.

Either way, for reasons that I can't quite remember, I decided having everything in one big transaction is best (plus, this way migrations are atomic). We still need to be able to calculate the index trait getters every runtime; to do this, every runtime, all transactions will be run; however, already-done transactions they will be aborted and all changes will be discarded except for database alterations, which will be used in the current runtime. A new parameter, 'dry: boolean', has been added to transaction running operations so that the user can avoid running impure code in transaction dry-runs.

Two big TODOs:
  1. Rewrite existing migraiton documentation
  2. Redesign the type registering system to fit with this new migration design. A type register action should be bound to a transaction, and should be rolled back on a transaction abort.