rework custom storable/indexable types 1. makes custom codecs be given as a part of database initialization instead of in migrations 2. removes all type checking regarding custom codecs
reason for (1): having custom codecs be a part of migration was misleading. it suggested that, for instance, Jinedb is able to use a codec from two versions ago (e.g. in order to upgrade the database to the current version). the truth is that Jinedb only has access to the most up-to-date codecs; this new design reflects that. also, requiring the codecs in the constructor make the data dependencies clearer than allowing them to be added after instantiation.
reason for (2): i think that typescript is simply not powerful enough to be able to handle this use-case well. there are too many factors at play: for instance, (1) that what is natively storable and what is natively indexable are two different types; or (2) that we need typescript to somehow know what types are custom-storable/custom-indexable. (the obvious solution is to require an e.g. `jineEncode` method on custom-* types, but the `jineDecode` would need to be on the class as a static method and i'm not sure if ts can detect that easily in this case). anyway, having types for this would be great, but it was being a huge difficulty so i decided to forgo it.